requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
drf-spectacular==0.25.1
grpcio==1.51.3
grpcio-tools==1.51.3
protobuf==4.22.0
pillow==9.4.0
psycopg2-binary==2.9.5
django-environ==0.10.0
gunicorn==20.1.0
django-allauth==0.53.1
